Moved all negative tests into a separate class NegativeTests
