+++ Blog.UI.Tests/Logs/Log.txt
@@ -668,3 +668,9 @@ System.NullReferenceException : Object reference not set to an instance of an ob
 [2017-05-30T12:10:50] [Passed] Register_Without_Password
 [2017-05-30T12:12:07] [Passed] Register_Without_Email
 [2017-05-30T12:18:47] [Passed] Register_Without_ConfirmPassword
+[2017-05-30T16:11:44] [Passed] Login_Blank_Email_Login_Unsuccessful
+[2017-05-30T16:11:46] [Passed] Login_InvalidPassword_LoginUnsuccessful
+[2017-05-30T16:11:47] [Passed] Register_Without_ConfirmPassword
+[2017-05-30T16:11:49] [Passed] Register_Without_Email
+[2017-05-30T16:11:50] [Passed] Register_Without_FullName
+[2017-05-30T16:11:52] [Passed] Register_Without_Password
