+++ Blog.UI.Tests/bin/Debug/nunit_random_seed.tmp
@@ -1,1 +1,1 @@
-1493316281
+296280046
